Split the 2,215-line conversations context test
ExUnit parallelises across modules and never within one, so a single module is
a floor on whichever partition draws it. conversations_context_test.exs was
2,215 lines and 160 tests in one module, and it took 29.4s — 28% of the
slowest partition's entire runtime. `--slowest` had already ruled out a few
pathological tests: its twelve slowest were 9.8% of its time. It was simply
one very large module.

Split along the section banners it already had, so every new module is the
context surface it was documenting anyway:

    conversations_context_test.exs      39 tests   sandboxes + conversations
    conversations_turns_test.exs        16         turns
    conversations_log_events_test.exs   42         log events, output, images
    conversations_wake_test.exs         17         wake_conversation/2
    conversations_start_test.exs        32         start_conversation/1 and /2
    conversations_read_state_test.exs   14         mark_read/2, list_conversations/2

Same 160 tests, 0 failures, and the six run in 7.8s where the one took 29.4s.
Across a full six-way partition sweep the slowest partition fell from 66.5s to
58.8s and the total from 252.5s to 237.8s.

Two things worth knowing for whoever tunes this next.

Partition assignment is by index over the sorted file list, not a stable hash
of each path: adding these five files moved almost every other file to a
different partition. Balance therefore cannot be hand-tuned — it reshuffles
whenever a test file is added or renamed, which is why CI's imbalance drifted
from 1.65x to 2.39x over a handful of unrelated merges.

And the bottleneck has moved. The slowest partition is now 48.4s of `async:
false` sync against 10.4s of async, with no dominant module — 140s of the
238s total is the serial phase, spread unevenly across 72 `async: false`
modules. Splitting more files will not touch that; reducing the number of
serial modules would.
