+++ apps/fountain/test/fountain/conversations_context_test.exs
@@ -1,11 +1,14 @@
 defmodule Fountain.ConversationsContextTest do
   use Fountain.DataCase, async: true
-  use Mimic
 
-  alias Fountain.{Agents, Conversations}
-  alias Fountain.Conversations.{Conversation, LogEvent, Sandbox, Turn}
+  alias Fountain.Conversations
+  alias Fountain.Conversations.{Conversation, Sandbox}
 
-  # ────────────────────────────────────────────────────────────────────────────
+  # Sandboxes and conversations: the CRUD surface of the context.
+  # Split out of the 2,215-line conversations_context_test.exs (#899): ExUnit
+  # parallelises across modules, never within one, so that single module was a
+  # 29.4s floor for whichever partition drew it.
+
   # Sandboxes
   # ────────────────────────────────────────────────────────────────────────────
 
@@ -397,1819 +400,4 @@ defmodule Fountain.ConversationsContextTest do
   end
 
   # ────────────────────────────────────────────────────────────────────────────
-  # Turns
-  # ────────────────────────────────────────────────────────────────────────────
-
-  describe "_unsafe_list_turns/1" do
-    test "returns empty list when conversation has no turns" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-
-      assert Conversations._unsafe_list_turns(conv.id) == []
-    end
-
-    test "returns all turns for the conversation" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      t1 = insert_turn(conv)
-      t2 = insert_turn(conv)
-
-      ids = Conversations._unsafe_list_turns(conv.id) |> Enum.map(& &1.id)
-      assert t1.id in ids
-      assert t2.id in ids
-    end
-
-    test "orders turns by turn_number ascending" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      t1 = insert_turn(conv)
-      t2 = insert_turn(conv)
-
-      [first, second] = Conversations._unsafe_list_turns(conv.id)
-      assert first.turn_number < second.turn_number
-      assert first.id == t1.id
-      assert second.id == t2.id
-    end
-
-    test "does not return turns from other conversations" do
-      user = insert_verified_user()
-      conv1 = insert_conversation(user_id: user.id)
-      conv2 = insert_conversation(user_id: user.id)
-      t1 = insert_turn(conv1)
-      _t2 = insert_turn(conv2)
-
-      results = Conversations._unsafe_list_turns(conv1.id)
-      assert length(results) == 1
-      assert hd(results).id == t1.id
-    end
-  end
-
-  describe "get_turn_by_conversation/3" do
-    test "returns the turn when turn_id, conversation_id and user_id match" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      result = Conversations.get_turn_by_conversation(turn.id, conv.id, user.id)
-      assert result.id == turn.id
-    end
-
-    test "returns nil when turn_id does not exist" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-
-      assert Conversations.get_turn_by_conversation(Ecto.UUID.generate(), conv.id, user.id) == nil
-    end
-
-    test "returns nil when turn belongs to a different conversation" do
-      user = insert_verified_user()
-      conv1 = insert_conversation(user_id: user.id)
-      conv2 = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv1)
-
-      assert Conversations.get_turn_by_conversation(turn.id, conv2.id, user.id) == nil
-    end
-
-    test "returns nil when the conversation belongs to a different user" do
-      owner = insert_verified_user()
-      other = insert_verified_user()
-      conv = insert_conversation(user_id: owner.id)
-      turn = insert_turn(conv)
-
-      assert Conversations.get_turn_by_conversation(turn.id, conv.id, other.id) == nil
-    end
-  end
-
-  describe "_unsafe_next_turn_number/1" do
-    test "returns 1 when conversation has no turns" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-
-      assert Conversations._unsafe_next_turn_number(conv.id) == 1
-    end
-
-    test "returns max_turn_number + 1 when turns exist" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      _t1 = insert_turn(conv)
-      _t2 = insert_turn(conv)
-
-      assert Conversations._unsafe_next_turn_number(conv.id) == 3
-    end
-
-    test "is not affected by turns from other conversations" do
-      user = insert_verified_user()
-      conv1 = insert_conversation(user_id: user.id)
-      conv2 = insert_conversation(user_id: user.id)
-      _t1 = insert_turn(conv1)
-      _t2 = insert_turn(conv1)
-      _t3 = insert_turn(conv1)
-
-      assert Conversations._unsafe_next_turn_number(conv2.id) == 1
-    end
-  end
-
-  describe "_unsafe_create_turn/1" do
-    test "creates a turn with valid attrs" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-
-      attrs = %{
-        conversation_id: conv.id,
-        turn_number: 1,
-        prompt: "Hello, world",
-        status: "pending"
-      }
-
-      assert {:ok, turn} = Conversations._unsafe_create_turn(attrs)
-      assert turn.conversation_id == conv.id
-      assert turn.turn_number == 1
-      assert turn.prompt == "Hello, world"
-      assert turn.status == "pending"
-    end
-
-    test "returns error changeset when required fields are missing" do
-      assert {:error, changeset} = Conversations._unsafe_create_turn(%{})
-      assert changeset.valid? == false
-      assert errors_on(changeset)[:conversation_id]
-      assert errors_on(changeset)[:turn_number]
-    end
-
-    test "returns error changeset when status is invalid" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-
-      attrs = %{
-        conversation_id: conv.id,
-        turn_number: 1,
-        prompt: "test",
-        status: "bogus"
-      }
-
-      assert {:error, changeset} = Conversations._unsafe_create_turn(attrs)
-      assert errors_on(changeset)[:status]
-    end
-  end
-
-  describe "_unsafe_update_turn/2" do
-    test "updates a turn with valid attrs" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      assert {:ok, updated} = Conversations._unsafe_update_turn(turn, %{status: "completed"})
-      assert updated.id == turn.id
-      assert updated.status == "completed"
-    end
-
-    test "returns error changeset when status is invalid" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      assert {:error, changeset} = Conversations._unsafe_update_turn(turn, %{status: "invalid"})
-      assert errors_on(changeset)[:status]
-    end
-  end
-
-  # ────────────────────────────────────────────────────────────────────────────
-  # Log events
-  # ────────────────────────────────────────────────────────────────────────────
-
-  describe "log!/1" do
-    test "inserts a LogEvent and returns the struct" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-
-      attrs = %{
-        conversation_id: conv.id,
-        kind: "output",
-        stream: "stdout",
-        data: "hello",
-        inserted_at: DateTime.utc_now()
-      }
-
-      event = Conversations.log!(attrs)
-      assert %LogEvent{} = event
-      assert is_integer(event.id)
-      assert event.conversation_id == conv.id
-      assert event.kind == "output"
-      assert event.data == "hello"
-    end
-
-    test "defaults inserted_at to current time when not provided" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-
-      attrs = %{
-        conversation_id: conv.id,
-        kind: "output"
-      }
-
-      before = DateTime.utc_now()
-      event = Conversations.log!(attrs)
-      after_time = DateTime.utc_now()
-
-      assert DateTime.compare(event.inserted_at, before) in [:gt, :eq]
-      assert DateTime.compare(event.inserted_at, after_time) in [:lt, :eq]
-    end
-
-    test "inserts a stage event" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-
-      event =
-        Conversations.log!(%{
-          conversation_id: conv.id,
-          kind: "stage",
-          stage: "provision",
-          state: "started",
-          inserted_at: DateTime.utc_now()
-        })
-
-      assert event.kind == "stage"
-      assert event.stage == "provision"
-      assert event.state == "started"
-    end
-  end
-
-  describe "_unsafe_list_log_events/3" do
-    setup do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      %{user: user, conv: conv}
-    end
-
-    test "returns all events when no opts given", %{conv: conv} do
-      e1 = insert_log_event(conv, kind: "output", stream: "stdout", data: "a")
-      e2 = insert_log_event(conv, kind: "output", stream: "stderr", data: "b")
-      e3 = insert_log_event(conv, kind: "stage", stream: "", stage: "provision")
-
-      ids = Conversations._unsafe_list_log_events(conv.id) |> Enum.map(& &1.id)
-      assert e1.id in ids
-      assert e2.id in ids
-      assert e3.id in ids
-    end
-
-    test "no filter with empty streams list returns all events", %{conv: conv} do
-      e1 = insert_log_event(conv, kind: "output", stream: "stdout")
-      e2 = insert_log_event(conv, kind: "stage", stream: "", stage: "provision")
-
-      ids = Conversations._unsafe_list_log_events(conv.id, 0, streams: []) |> Enum.map(& &1.id)
-      assert e1.id in ids
-      assert e2.id in ids
-    end
-
-    test "streams: [\"stdout\"] returns only stdout events", %{conv: conv} do
-      stdout = insert_log_event(conv, kind: "output", stream: "stdout", data: "out")
-      stderr = insert_log_event(conv, kind: "output", stream: "stderr", data: "err")
-      stage = insert_log_event(conv, kind: "stage", stream: "", stage: "provision")
-
-      results = Conversations._unsafe_list_log_events(conv.id, 0, streams: ["stdout"])
-      ids = Enum.map(results, & &1.id)
-      assert stdout.id in ids
-      refute stderr.id in ids
-      refute stage.id in ids
-    end
-
-    test "streams: [\"stderr\"] returns only stderr events", %{conv: conv} do
-      stdout = insert_log_event(conv, kind: "output", stream: "stdout")
-      stderr = insert_log_event(conv, kind: "output", stream: "stderr")
-      stage = insert_log_event(conv, kind: "stage", stream: "", stage: "provision")
-
-      results = Conversations._unsafe_list_log_events(conv.id, 0, streams: ["stderr"])
-      ids = Enum.map(results, & &1.id)
-      refute stdout.id in ids
-      assert stderr.id in ids
-      refute stage.id in ids
-    end
-
-    test "streams: [\"stage\"] returns only stage kind events", %{conv: conv} do
-      stdout = insert_log_event(conv, kind: "output", stream: "stdout")
-      stderr = insert_log_event(conv, kind: "output", stream: "stderr")
-      stage = insert_log_event(conv, kind: "stage", stream: "", stage: "provision")
-
-      results = Conversations._unsafe_list_log_events(conv.id, 0, streams: ["stage"])
-      ids = Enum.map(results, & &1.id)
-      refute stdout.id in ids
-      refute stderr.id in ids
-      assert stage.id in ids
-    end
-
-    test ~s|streams: ["stdout", "stage"] returns stdout and stage events|, %{conv: conv} do
-      stdout = insert_log_event(conv, kind: "output", stream: "stdout")
-      stderr = insert_log_event(conv, kind: "output", stream: "stderr")
-      stage = insert_log_event(conv, kind: "stage", stream: "", stage: "provision")
-
-      results = Conversations._unsafe_list_log_events(conv.id, 0, streams: ["stdout", "stage"])
-      ids = Enum.map(results, & &1.id)
-      assert stdout.id in ids
-      refute stderr.id in ids
-      assert stage.id in ids
-    end
-
-    test "streams: [\"unknown\"] returns empty list (unknown stream filter)", %{conv: conv} do
-      _e = insert_log_event(conv, kind: "output", stream: "stdout")
-
-      assert Conversations._unsafe_list_log_events(conv.id, 0, streams: ["unknown"]) == []
-    end
-
-    test "after_id filters events with id greater than after_id", %{conv: conv} do
-      e1 = insert_log_event(conv, kind: "output", stream: "stdout")
-      e2 = insert_log_event(conv, kind: "output", stream: "stdout")
-      e3 = insert_log_event(conv, kind: "output", stream: "stdout")
-
-      ids = Conversations._unsafe_list_log_events(conv.id, e1.id) |> Enum.map(& &1.id)
-      refute e1.id in ids
-      assert e2.id in ids
-      assert e3.id in ids
-    end
-
-    test "returns events ordered by id ascending", %{conv: conv} do
-      e1 = insert_log_event(conv, kind: "output", stream: "stdout")
-      e2 = insert_log_event(conv, kind: "output", stream: "stdout")
-      e3 = insert_log_event(conv, kind: "output", stream: "stdout")
-
-      ids = Conversations._unsafe_list_log_events(conv.id) |> Enum.map(& &1.id)
-      assert ids == Enum.sort(ids)
-      assert hd(ids) == e1.id
-      assert List.last(ids) == e3.id
-    end
-
-    test "does not return events from other conversations", %{conv: conv} do
-      user = insert_verified_user()
-      other_conv = insert_conversation(user_id: user.id)
-      _other = insert_log_event(other_conv, kind: "output", stream: "stdout")
-      mine = insert_log_event(conv, kind: "output", stream: "stdout")
-
-      results = Conversations._unsafe_list_log_events(conv.id)
-      ids = Enum.map(results, & &1.id)
-      assert ids == [mine.id]
-    end
-  end
-
-  # ────────────────────────────────────────────────────────────────────────────
-  # _unsafe_list_sandboxes_admin/0
-  # ────────────────────────────────────────────────────────────────────────────
-
-  describe "_unsafe_list_sandboxes_admin/0" do
-    test "returns sandboxes with pending and ready statuses" do
-      user = insert_verified_user()
-      pending = insert_sandbox(user_id: user.id)
-      ready = insert_sandbox(user_id: user.id)
-      {:ok, _} = Conversations.update_sandbox(ready, %{status: "ready"})
-
-      ids = Conversations._unsafe_list_sandboxes_admin() |> Enum.map(& &1.id)
-      assert pending.id in ids
-      assert ready.id in ids
-    end
-
-    test "excludes sandboxes with terminated status" do
-      user = insert_verified_user()
-      sandbox = insert_sandbox(user_id: user.id)
-      {:ok, _} = Conversations.update_sandbox(sandbox, %{status: "terminated"})
-
-      ids = Conversations._unsafe_list_sandboxes_admin() |> Enum.map(& &1.id)
-      refute sandbox.id in ids
-    end
-
-    test "excludes sandboxes with failed status" do
-      user = insert_verified_user()
-      sandbox = insert_sandbox(user_id: user.id)
-      {:ok, _} = Conversations.update_sandbox(sandbox, %{status: "failed"})
-
-      ids = Conversations._unsafe_list_sandboxes_admin() |> Enum.map(& &1.id)
-      refute sandbox.id in ids
-    end
-
-    test "includes pending but not terminated when both exist" do
-      user = insert_verified_user()
-      active = insert_sandbox(user_id: user.id)
-      terminated = insert_sandbox(user_id: user.id)
-      {:ok, _} = Conversations.update_sandbox(terminated, %{status: "terminated"})
-
-      ids = Conversations._unsafe_list_sandboxes_admin() |> Enum.map(& &1.id)
-      assert active.id in ids
-      refute terminated.id in ids
-    end
-
-    test "preloads user association" do
-      user = insert_verified_user()
-      _sandbox = insert_sandbox(user_id: user.id)
-
-      results = Conversations._unsafe_list_sandboxes_admin()
-      assert results != []
-      result = Enum.find(results, &(&1.user_id == user.id))
-      assert result.user.id == user.id
-    end
-
-    test "preloads conversations association" do
-      user = insert_verified_user()
-      sandbox = insert_sandbox(user_id: user.id)
-      _conv = insert_conversation(user_id: user.id, sandbox_id: sandbox.id)
-
-      results = Conversations._unsafe_list_sandboxes_admin()
-      result = Enum.find(results, &(&1.id == sandbox.id))
-      assert is_list(result.conversations)
-    end
-
-    test "returns empty list when all sandboxes are in terminal states" do
-      user = insert_verified_user()
-      s1 = insert_sandbox(user_id: user.id)
-      s2 = insert_sandbox(user_id: user.id)
-      {:ok, _} = Conversations.update_sandbox(s1, %{status: "terminated"})
-      {:ok, _} = Conversations.update_sandbox(s2, %{status: "failed"})
-
-      # All sandboxes in this test's db partition are terminal
-      results = Conversations._unsafe_list_sandboxes_admin()
-      ids = Enum.map(results, & &1.id)
-      refute s1.id in ids
-      refute s2.id in ids
-    end
-  end
-
-  # ────────────────────────────────────────────────────────────────────────────
-  # _unsafe_list_resumable_conversations/0
-  # ────────────────────────────────────────────────────────────────────────────
-
-  describe "_unsafe_list_resumable_conversations/0" do
-    test "returns idle conversation whose sandbox is ready" do
-      user = insert_verified_user()
-      sandbox = insert_sandbox(user_id: user.id)
-      {:ok, _} = Conversations.update_sandbox(sandbox, %{status: "ready"})
-      conv = insert_conversation(user_id: user.id, status: "idle", sandbox_id: sandbox.id)
-
-      ids = Conversations._unsafe_list_resumable_conversations() |> Enum.map(& &1.id)
-      assert conv.id in ids
-    end
-
-    test "returns running conversation whose sandbox is ready" do
-      user = insert_verified_user()
-      sandbox = insert_sandbox(user_id: user.id)
-      {:ok, _} = Conversations.update_sandbox(sandbox, %{status: "ready"})
-      conv = insert_conversation(user_id: user.id, status: "running", sandbox_id: sandbox.id)
-
-      ids = Conversations._unsafe_list_resumable_conversations() |> Enum.map(& &1.id)
-      assert conv.id in ids
-    end
-
-    test "excludes idle conversation whose sandbox is not ready (pending)" do
-      user = insert_verified_user()
-      sandbox = insert_sandbox(user_id: user.id)
-      # sandbox remains "pending"
-      conv = insert_conversation(user_id: user.id, status: "idle", sandbox_id: sandbox.id)
-
-      ids = Conversations._unsafe_list_resumable_conversations() |> Enum.map(& &1.id)
-      refute conv.id in ids
-    end
-
-    test "excludes terminated conversation even when sandbox is ready" do
-      user = insert_verified_user()
-      sandbox = insert_sandbox(user_id: user.id)
-      {:ok, _} = Conversations.update_sandbox(sandbox, %{status: "ready"})
-      conv = insert_conversation(user_id: user.id, status: "terminated", sandbox_id: sandbox.id)
-
-      ids = Conversations._unsafe_list_resumable_conversations() |> Enum.map(& &1.id)
-      refute conv.id in ids
-    end
-
-    test "excludes a terminated conversation even when sandbox is ready" do
-      # The sandbox being usable is not enough; the conversation has to be one
-      # someone can still talk to. This used to be asserted with "completed",
-      # a status nothing could produce, so it proved nothing.
-      user = insert_verified_user()
-      sandbox = insert_sandbox(user_id: user.id)
-      {:ok, _} = Conversations.update_sandbox(sandbox, %{status: "ready"})
-      conv = insert_conversation(user_id: user.id, status: "terminated", sandbox_id: sandbox.id)
-
-      ids = Conversations._unsafe_list_resumable_conversations() |> Enum.map(& &1.id)
-      refute conv.id in ids
-    end
-
-    test "preloads sandbox association" do
-      user = insert_verified_user()
-      sandbox = insert_sandbox(user_id: user.id)
-      {:ok, _} = Conversations.update_sandbox(sandbox, %{status: "ready"})
-      conv = insert_conversation(user_id: user.id, status: "idle", sandbox_id: sandbox.id)
-
-      results = Conversations._unsafe_list_resumable_conversations()
-      result = Enum.find(results, &(&1.id == conv.id))
-      assert %Sandbox{} = result.sandbox
-      assert result.sandbox.id == sandbox.id
-    end
-
-    test "returns empty list when no resumable conversations exist" do
-      assert Conversations._unsafe_list_resumable_conversations() == []
-    end
-  end
-
-  describe "_unsafe_output_bytes_by_stream/2" do
-    test "returns empty map when there are no output events for the turn" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      assert Conversations._unsafe_output_bytes_by_stream(conv.id, turn.id) == %{}
-    end
-
-    test "sums byte lengths of data grouped by stream" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      # "hello" = 5 bytes, "world" = 5 bytes → stdout total = 10
-      insert_log_event(conv,
-        kind: "output",
-        stream: "stdout",
-        data: "hello",
-        turn_id: turn.id
-      )
-
-      insert_log_event(conv,
-        kind: "output",
-        stream: "stdout",
-        data: "world",
-        turn_id: turn.id
-      )
-
-      # "err" = 3 bytes → stderr total = 3
-      insert_log_event(conv,
-        kind: "output",
-        stream: "stderr",
-        data: "err",
-        turn_id: turn.id
-      )
-
-      result = Conversations._unsafe_output_bytes_by_stream(conv.id, turn.id)
-      assert result["stdout"] == 10
-      assert result["stderr"] == 3
-    end
-
-    test "excludes non-output kind events (stage events)" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      insert_log_event(conv,
-        kind: "stage",
-        stage: "provision",
-        data: "stage-data",
-        turn_id: turn.id
-      )
-
-      result = Conversations._unsafe_output_bytes_by_stream(conv.id, turn.id)
-      # Stage events have empty stream and are kind "stage", not "output"
-      assert result == %{}
-    end
-
-    test "excludes events from other turns" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn1 = insert_turn(conv)
-      turn2 = insert_turn(conv)
-
-      insert_log_event(conv,
-        kind: "output",
-        stream: "stdout",
-        data: "from-turn-2",
-        turn_id: turn2.id
-      )
-
-      result = Conversations._unsafe_output_bytes_by_stream(conv.id, turn1.id)
-      assert result == %{}
-    end
-
-    test "excludes events from other conversations" do
-      user = insert_verified_user()
-      conv1 = insert_conversation(user_id: user.id)
-      conv2 = insert_conversation(user_id: user.id)
-      turn1 = insert_turn(conv1)
-      turn2 = insert_turn(conv2)
-
-      insert_log_event(conv2,
-        kind: "output",
-        stream: "stdout",
-        data: "from-other-conv",
-        turn_id: turn2.id
-      )
-
-      result = Conversations._unsafe_output_bytes_by_stream(conv1.id, turn1.id)
-      assert result == %{}
-    end
-
-    test "returns map with stream keys as strings" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      insert_log_event(conv,
-        kind: "output",
-        stream: "stdout",
-        data: "abc",
-        turn_id: turn.id
-      )
-
-      result = Conversations._unsafe_output_bytes_by_stream(conv.id, turn.id)
-      assert Map.has_key?(result, "stdout")
-    end
-  end
-
-  # ────────────────────────────────────────────────────────────────────────────
-  # _unsafe_list_conversations_by_activity/0
-  # ────────────────────────────────────────────────────────────────────────────
-
-  # ────────────────────────────────────────────────────────────────────────────
-  # _unsafe_insert_turn_images/2 and _unsafe_get_turn_image/2
-  # ────────────────────────────────────────────────────────────────────────────
-
-  # Helper to insert a TurnImage via changeset directly, for tests that need a
-  # row without exercising _unsafe_insert_turn_images/2.
-  defp insert_turn_image!(turn_id, position, media_type, data) do
-    %Fountain.Conversations.TurnImage{}
-    |> Fountain.Conversations.TurnImage.changeset(%{
-      turn_id: turn_id,
-      position: position,
-      media_type: media_type,
-      data: data
-    })
-    |> Ecto.Changeset.put_change(:inserted_at, DateTime.utc_now() |> DateTime.truncate(:second))
-    |> Repo.insert!()
-  end
-
-  describe "_unsafe_insert_turn_images/2" do
-    test "returns {:ok, 0} immediately when images list is empty" do
-      assert {:ok, 0} = Conversations._unsafe_insert_turn_images(Ecto.UUID.generate(), [])
-    end
-
-    test "inserts images and returns count" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-      images = [%{media_type: "image/png", data: <<1, 2, 3>>}]
-      assert {:ok, 1} = Conversations._unsafe_insert_turn_images(turn.id, images)
-    end
-
-    test "positions are assigned in order" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      assert {:ok, 2} =
-               Conversations._unsafe_insert_turn_images(turn.id, [
-                 %{media_type: "image/png", data: <<1>>},
-                 %{media_type: "image/jpeg", data: <<2>>}
-               ])
-
-      assert Conversations._unsafe_get_turn_image(turn.id, 0).media_type == "image/png"
-      assert Conversations._unsafe_get_turn_image(turn.id, 1).media_type == "image/jpeg"
-    end
-
-    test "a disallowed media type is refused instead of stored" do
-      # This used to go through Repo.insert_all against a raw table name, so the
-      # schema's allowlist never ran and the schema described validation that
-      # nothing performed.
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      assert {:error, changeset} =
-               Conversations._unsafe_insert_turn_images(turn.id, [
-                 %{media_type: "text/html", data: "<script>alert(1)</script>"}
-               ])
-
-      assert {"is invalid", _} = changeset.errors[:media_type]
-      assert Conversations._unsafe_get_turn_image(turn.id, 0) == nil
-    end
-
-    test "an unknown turn is refused rather than orphaning a row" do
-      assert {:error, changeset} =
-               Conversations._unsafe_insert_turn_images(Ecto.UUID.generate(), [
-                 %{media_type: "image/png", data: <<1>>}
-               ])
-
-      refute changeset.valid? and changeset.errors == []
-    end
-  end
-
-  describe "_unsafe_get_turn_image/2" do
-    test "returns the TurnImage when turn_id and position match" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      insert_turn_image!(turn.id, 0, "image/png", <<7, 8, 9>>)
-
-      result = Conversations._unsafe_get_turn_image(turn.id, 0)
-      assert result != nil
-      assert result.turn_id == turn.id
-      assert result.position == 0
-      assert result.media_type == "image/png"
-    end
-
-    test "returns nil when no image exists at the given position" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      assert Conversations._unsafe_get_turn_image(turn.id, 99) == nil
-    end
-
-    test "returns nil when turn_id does not exist" do
-      assert Conversations._unsafe_get_turn_image(Ecto.UUID.generate(), 0) == nil
-    end
-
-    test "returns nil when position belongs to a different turn" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn1 = insert_turn(conv)
-      turn2 = insert_turn(conv)
-
-      insert_turn_image!(turn1.id, 0, "image/png", <<1>>)
-
-      # turn1 has an image at position 0, but turn2 does not
-      assert Conversations._unsafe_get_turn_image(turn2.id, 0) == nil
-    end
-  end
-
-  # ────────────────────────────────────────────────────────────────────────────
-  # wake_conversation/2
-  # ────────────────────────────────────────────────────────────────────────────
-
-  describe "wake_conversation/2" do
-    test "returns {:error, :not_found} when conversation does not exist" do
-      assert {:error, :not_found} = Conversations.wake_conversation(Ecto.UUID.generate())
-    end
-
-    test "returns {:error, :gone} when conversation status is terminated" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id, status: "terminated")
-
-      assert {:error, :gone} = Conversations.wake_conversation(conv.id)
-    end
-
-    test "returns {:error, :gone} when conversation status is failed" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id, status: "failed")
-
-      assert {:error, :gone} = Conversations.wake_conversation(conv.id)
-    end
-
-    test "an idle conversation whose sandbox was reclaimed is still resumable" do
-      # The case #167 created and the reason `completed` was never needed:
-      # reclaiming an idle sandbox leaves the conversation resumable rather
-      # than closing it, so waking it must not be refused as :gone.
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      sandbox = insert_sandbox(user_id: user.id, status: "terminated")
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      refute match?({:error, :gone}, Conversations.wake_conversation(conv.id))
-    end
-
-    test "returns {:error, :no_agent} when conversation has no agent_id" do
-      user = insert_verified_user()
-      # insert_conversation does not set an agent by default, so agent_id is nil
-      conv = insert_conversation(user_id: user.id, status: "idle")
-
-      assert {:error, :no_agent} = Conversations.wake_conversation(conv.id)
-    end
-
-    test "returns {:ok, conv} reusing existing sandbox when sprite is still alive" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      sandbox = insert_sandbox(user_id: user.id, sprite_name: "test-sprite-alive")
-      {:ok, sandbox} = Conversations.update_sandbox(sandbox, %{status: "ready"})
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      fake_client = %{}
-
-      stub(Fountain.Sandbox.Sprites, :get, fn _handle ->
-        {:ok, %{status: :running, raw: %{name: "test-sprite-alive"}}}
-      end)
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _supervisor, _child_spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      assert {:ok, _conv} = Conversations.wake_conversation(conv.id)
-    end
-
-    test "a reuse that loses the start race hands the prompt to the winner (#667)" do
-      # Two concurrent wakes of the same suspended/ready sandbox: the loser
-      # used to get the raw {:error, {:already_started, pid}} back from
-      # start_conversation_server, so the prompt was silently dropped instead
-      # of reaching the winner's server.
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      sandbox = insert_sandbox(user_id: user.id, sprite_name: "test-sprite-race")
-      {:ok, sandbox} = Conversations.update_sandbox(sandbox, %{status: "ready"})
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      winner = spawn(fn -> Process.sleep(:infinity) end)
-      test_pid = self()
-
-      stub(Fountain.Sandbox.Sprites, :get, fn _handle ->
-        {:ok, %{status: :running, raw: %{name: "test-sprite-race"}}}
-      end)
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _supervisor, _child_spec ->
-        {:error, {:already_started, winner}}
-      end)
-
-      stub(Fountain.Conversations.ConversationServer, :queue_initial_prompt, fn pid, prompt ->
-        send(test_pid, {:queued, pid, prompt})
-        :ok
-      end)
-
-      assert {:ok, woken} = Conversations.wake_conversation(conv.id, "hello")
-      assert_receive {:queued, ^winner, "hello"}
-
-      # Reuse touches no row — the conversation still names the sandbox it
-      # started with.
-      assert woken.sandbox_id == sandbox.id
-    end
-
-    test "waking a suspended sandbox flips it to ready and stamps the clock" do
-      # The core of decisions/0017: the parked sprite is reused, re-added to
-      # the quota, and the max-lifetime ceiling restarts from the wake.
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      sandbox = insert_sandbox(user_id: user.id, sprite_name: "test-sprite-parked")
-      {:ok, sandbox} = Conversations.update_sandbox(sandbox, %{status: "suspended"})
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      stub(Fountain.Sandbox.Sprites, :get, fn _handle ->
-        {:ok, %{status: :suspended, raw: %{name: "test-sprite-parked"}}}
-      end)
-
-      # resume/1 is the explicit wake call (a probe on Sprites); the adapter
-      # is the stubbing seam, so it needs its own stub here.
-      stub(Fountain.Sandbox.Sprites, :resume, fn handle -> {:ok, handle} end)
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _supervisor, _child_spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      assert {:ok, woken} = Conversations.wake_conversation(conv.id)
-      # Same sandbox — no fresh sprite was provisioned.
-      assert woken.sandbox_id == sandbox.id
-
-      reloaded = Repo.reload(sandbox)
-      assert reloaded.status == "ready"
-      assert %DateTime{} = reloaded.last_resumed_at
-    end
-
-    test "waking a suspended sandbox re-runs the quota gate" do
-      # A parked sprite is free; waking it is compute again. A user at their
-      # cap must be refused, exactly as if they were starting a conversation.
-      user = insert_verified_user()
-      {:ok, user} = Fountain.Accounts.update_sandbox_limit(user, 1)
-      agent = insert_agent(user_id: user.id)
-
-      insert_sandbox(user_id: user.id, status: "ready")
-
-      sandbox = insert_sandbox(user_id: user.id, sprite_name: "test-sprite-capped")
-      {:ok, sandbox} = Conversations.update_sandbox(sandbox, %{status: "suspended"})
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      stub(Fountain.Sandbox.Sprites, :get, fn _handle -> {:ok, %{status: :unknown, raw: %{}}} end)
-
-      assert {:error, {:sandbox_quota_exceeded, _}} = Conversations.wake_conversation(conv.id)
-      # Refused means still parked — the row must not be half-woken.
-      assert Repo.reload(sandbox).status == "suspended"
-    end
-
-    test "a transient sprite probe failure does not give up a suspended sandbox" do
-      # Falling to :create_new would retire the row and route the still-live
-      # sprite — the agent's memory — to the reaper's destroy pass. Only a
-      # definitive not-found may do that; anything else fails retryably.
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      sandbox = insert_sandbox(user_id: user.id, sprite_name: "test-sprite-blip")
-      {:ok, sandbox} = Conversations.update_sandbox(sandbox, %{status: "suspended"})
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      stub(Fountain.Sandbox.Sprites, :get, fn _handle -> {:error, {:unavailable, :timeout}} end)
-
-      assert {:error, :sprite_probe_failed} = Conversations.wake_conversation(conv.id)
-      assert Repo.reload(sandbox).status == "suspended"
-    end
-
-    test "a transient sprite probe failure does not give up a ready sandbox either (#799)" do
-      # A `ready` row whose server is gone (deploy, crash, partition) is the
-      # same parked disk as a suspended one. Until #799 only `suspended` was
-      # protected; a `ready` row fell to :create_new, was marked terminated,
-      # and the reaper destroyed a sprite that was fine.
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      sandbox = insert_sandbox(user_id: user.id, sprite_name: "test-sprite-blip-ready")
-      {:ok, sandbox} = Conversations.update_sandbox(sandbox, %{status: "ready"})
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      stub(Fountain.Sandbox.Sprites, :get, fn _handle ->
-        {:error, {:unavailable, %Req.TransportError{reason: :nxdomain}}}
-      end)
-
-      reject(&Horde.DynamicSupervisor.start_child/2)
-
-      assert {:error, :sprite_probe_failed} = Conversations.wake_conversation(conv.id)
-      reloaded = Repo.reload(sandbox)
-      assert reloaded.status == "ready"
-      assert is_nil(reloaded.terminated_at)
-      assert Repo.reload(conv).sandbox_id == sandbox.id
-    end
-
-    test "a definitively gone sprite retires the suspended sandbox and provisions fresh" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      sandbox = insert_sandbox(user_id: user.id, sprite_name: "test-sprite-vanished")
-      {:ok, sandbox} = Conversations.update_sandbox(sandbox, %{status: "suspended"})
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      stub(Fountain.Sandbox.Sprites, :get, fn _handle -> {:error, :not_found} end)
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _supervisor, _child_spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      assert {:ok, woken} = Conversations.wake_conversation(conv.id)
-      assert woken.sandbox_id != sandbox.id
-      assert Repo.reload(sandbox).status == "terminated"
-    end
-
-    test "returns {:ok, conv} creating fresh sandbox when sprite is gone" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      sandbox = insert_sandbox(user_id: user.id, sprite_name: "test-sprite-gone")
-      {:ok, sandbox} = Conversations.update_sandbox(sandbox, %{status: "ready"})
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      fake_client = %{}
-
-      stub(Fountain.Sandbox.Sprites, :get, fn _handle -> {:error, :not_found} end)
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _supervisor, _child_spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      assert {:ok, _conv} = Conversations.wake_conversation(conv.id)
-    end
-
-    test "returns {:ok, conv} creating fresh sandbox when sandbox is pending (not ready)" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-
-      # A `pending` sandbox with no server anywhere: the provision died with
-      # its BEAM. After the registry settle window (#800) the wake gives up
-      # waiting and provisions fresh.
-      sandbox = insert_sandbox(user_id: user.id)
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _supervisor, _child_spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      assert {:ok, woken} = Conversations.wake_conversation(conv.id)
-      assert woken.sandbox_id != sandbox.id
-    end
-
-    test "a pending sandbox whose server appears during the settle window is handed the prompt, not raced (#800)" do
-      # `session/new` (POST /api/conversations) starts the server via Horde,
-      # which may place it on another pod; the first prompt arrives ~30 ms
-      # later on this pod, sees a `pending` row and — before #800 — missed
-      # the registry and took :create_new: two servers, two sprites, ~21 s of
-      # provisioning each, and a name conflict that killed the loser after
-      # the fact, leaving an orphan `ready` row. The prompt path now waits
-      # for the registry to catch up and hands the prompt to the server it
-      # finds.
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      sandbox = insert_sandbox(user_id: user.id)
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      test = self()
-
-      # A stand-in for the first server: registers under the conversation's
-      # name a beat after the wake starts looking, then relays the prompt
-      # cast it receives.
-      late_server =
-        spawn_link(fn ->
-          Process.sleep(40)
-          {:ok, _} = Horde.Registry.register(Fountain.ConversationRegistry, conv.id, nil)
-          send(test, :registered)
-
-          receive do
-            {:"$gen_cast", {:initial_prompt, prompt, images}} ->
-              send(test, {:handed_off, prompt, images})
-          end
-        end)
-
-      # No second server, no second sandbox.
-      reject(&Horde.DynamicSupervisor.start_child/2)
-
-      assert {:ok, woken} = Conversations.wake_conversation(conv.id, "hello")
-      assert_receive :registered
-      assert_receive {:handed_off, "hello", []}, 500
-
-      assert woken.sandbox_id == sandbox.id
-      assert Repo.reload(sandbox).status == "pending"
-
-      assert Repo.aggregate(
-               from(s in Fountain.Conversations.Sandbox, where: s.user_id == ^user.id),
-               :count
-             ) == 1
-
-      Process.exit(late_server, :kill)
-    end
-
-    test "returns {:ok, conv} when old sandbox is already terminated (mark_old_sandbox_terminated no-op)" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      sandbox = insert_sandbox(user_id: user.id, sprite_name: "test-sprite-terminated")
-      {:ok, sandbox} = Conversations.update_sandbox(sandbox, %{status: "terminated"})
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _supervisor, _child_spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      # sandbox status is "terminated" (not "ready"), so maybe_reuse_sandbox returns :create_new,
-      # which calls create_fresh_sandbox_and_start -> mark_old_sandbox_terminated(sandbox.id)
-      # where the sandbox is already terminated, hitting the no-op branch
-      assert {:ok, _conv} = Conversations.wake_conversation(conv.id)
-    end
-
-    test "mark_old_sandbox_terminated handles deleted sandbox gracefully" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      # sandbox remains "pending" so maybe_reuse_sandbox returns :create_new
-      sandbox = insert_sandbox(user_id: user.id)
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      # Point the conversation at a non-existent sandbox_id (bypassing FK) so that
-      # mark_old_sandbox_terminated receives an id whose get_sandbox returns nil,
-      # hitting the nil -> :ok branch (line 635).
-      ghost_sandbox_id = Ecto.UUID.generate()
-      {:ok, ghost_uuid_bin} = Ecto.UUID.dump(ghost_sandbox_id)
-      {:ok, conv_id_bin} = Ecto.UUID.dump(conv.id)
-
-      # Temporarily disable FK checks, update the conversation to reference a
-      # non-existent sandbox_id, then re-enable. This simulates the case where
-      # a sandbox was deleted out-of-band (e.g., admin cleanup) so that
-      # mark_old_sandbox_terminated receives an id whose get_sandbox returns nil.
-      Ecto.Adapters.SQL.query!(Fountain.Repo, "SET session_replication_role = replica", [])
-
-      Ecto.Adapters.SQL.query!(
-        Fountain.Repo,
-        "UPDATE conversations SET sandbox_id = $1 WHERE id = $2",
-        [ghost_uuid_bin, conv_id_bin]
-      )
-
-      # Also delete the original sandbox record now that the FK is no longer referenced.
-      Ecto.Adapters.SQL.query!(Fountain.Repo, "SET session_replication_role = DEFAULT", [])
-      Fountain.Repo.delete!(sandbox)
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _supervisor, _child_spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      assert {:ok, _conv} = Conversations.wake_conversation(conv.id)
-    end
-  end
-
-  # ────────────────────────────────────────────────────────────────────────────
-  # start_conversation/1
-  # ────────────────────────────────────────────────────────────────────────────
-
-  describe "wake_conversation/2 — provider resume" do
-    test "a failed resume leaves the row suspended and fails retryably" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      sandbox = insert_sandbox(user_id: user.id, sprite_name: "parked-wont-wake")
-      {:ok, sandbox} = Conversations.update_sandbox(sandbox, %{status: "suspended"})
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      stub(Fountain.Sandbox.Sprites, :get, fn _handle ->
-        {:ok, %{status: :suspended, raw: %{}}}
-      end)
-
-      stub(Fountain.Sandbox.Sprites, :resume, fn _handle ->
-        {:error, {:unavailable, :timeout}}
-      end)
-
-      assert {:error, :sandbox_resume_failed} = Conversations.wake_conversation(conv.id)
-      assert Repo.reload(sandbox).status == "suspended"
-    end
-  end
-
-  describe "wake_conversation/2 — provider stickiness" do
-    test "a suspended sandbox on a disabled provider refuses to wake rather than retire" do
-      # Falling through to :create_new would retire the row and orphan (or
-      # lose) the parked disk; a retryable error keeps the agent's memory
-      # safe until the operator restores the provider's credentials.
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      sandbox = insert_sandbox(user_id: user.id, sprite_name: "parked-on-e2b")
-
-      {:ok, sandbox} =
-        Conversations.update_sandbox(sandbox, %{status: "suspended", provider: "e2b"})
-
-      conv = insert_conversation(user_id: user.id, agent: agent, sandbox: sandbox, status: "idle")
-
-      assert {:error, {:sandbox_provider_disabled, :e2b}} =
-               Conversations.wake_conversation(conv.id)
-
-      assert Repo.reload(sandbox).status == "suspended"
-    end
-  end
-
-  describe "start_conversation/1" do
-    test "the sandbox row is stamped with the resolved provider" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      stub(Horde.DynamicSupervisor, :start_child, fn _s, _spec -> {:ok, spawn(fn -> :ok end)} end)
-
-      assert {:ok, conv} =
-               Conversations.start_conversation(%{"agent_id" => agent.id, "user_id" => user.id})
-
-      sandbox = Conversations._unsafe_get_sandbox!(conv.sandbox_id)
-      assert sandbox.provider == "sprites"
-    end
-
-    test "an agent pinned to a disabled provider is refused before any row is allocated" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-
-      # Simulate drift: the override was saved while the provider was
-      # configured, and its credentials/adapter have since gone away. The
-      # changeset guards saves, so write the column directly.
-      {1, _} =
-        Repo.update_all(
-          Ecto.Query.from(a in Fountain.Agents.Agent, where: a.id == ^agent.id),
-          set: [sandbox_provider: "e2b"]
-        )
-
-      before = Fountain.Quotas.active_sandbox_count(user.id)
-
-      assert {:error, {:sandbox_provider_disabled, :e2b}} =
-               Conversations.start_conversation(%{"agent_id" => agent.id, "user_id" => user.id})
-
-      assert Fountain.Quotas.active_sandbox_count(user.id) == before
-    end
-
-    test "returns {:error, :vault_not_found} when vault_id belongs to a different user" do
-      user1 = insert_verified_user()
-      user2 = insert_verified_user()
-      agent = insert_agent(user_id: user1.id)
-      # vault belongs to user2, not user1
-      vault = insert_vault(user_id: user2.id)
-
-      attrs = %{
-        "agent_id" => agent.id,
-        "user_id" => user1.id,
-        "vault_id" => vault.id
-      }
-
-      assert {:error, :vault_not_found} = Conversations.start_conversation(attrs)
-    end
-
-    test "is refused once the tenant is at its concurrent-sandbox cap" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-
-      for _ <- 1..Fountain.Quotas.default_limit(),
-          do: insert_sandbox(user_id: user.id, status: "ready")
-
-      assert {:error, {:sandbox_quota_exceeded, %{count: 5, limit: 5}}} =
-               Conversations.start_conversation(%{"agent_id" => agent.id, "user_id" => user.id})
-    end
-
-    test "the cap is checked before any sandbox row is allocated" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      for _ <- 1..5, do: insert_sandbox(user_id: user.id, status: "ready")
-
-      before = Fountain.Quotas.active_sandbox_count(user.id)
-
-      assert {:error, _} =
-               Conversations.start_conversation(%{"agent_id" => agent.id, "user_id" => user.id})
-
-      # A denial that still allocated would let a caller ratchet past the cap
-      # by retrying, which is the failure mode the cap exists to prevent.
-      assert Fountain.Quotas.active_sandbox_count(user.id) == before
-    end
-
-    test "one tenant at its cap does not block another" do
-      capped = insert_verified_user()
-      other = insert_verified_user()
-      for _ <- 1..5, do: insert_sandbox(user_id: capped.id, status: "ready")
-      agent = insert_agent(user_id: other.id)
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _s, _spec -> {:ok, spawn(fn -> :ok end)} end)
-
-      assert {:ok, _} =
-               Conversations.start_conversation(%{"agent_id" => agent.id, "user_id" => other.id})
-    end
-
-    test "terminated sandboxes free capacity" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      [first | _] = for _ <- 1..5, do: insert_sandbox(user_id: user.id, status: "ready")
-
-      assert {:error, _} =
-               Conversations.start_conversation(%{"agent_id" => agent.id, "user_id" => user.id})
-
-      {:ok, _} = Conversations.update_sandbox(first, %{status: "terminated"})
-      stub(Horde.DynamicSupervisor, :start_child, fn _s, _spec -> {:ok, spawn(fn -> :ok end)} end)
-
-      assert {:ok, _} =
-               Conversations.start_conversation(%{"agent_id" => agent.id, "user_id" => user.id})
-    end
-
-    test "creates sandbox, conversation, and starts server", %{} do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _supervisor, _child_spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      attrs = %{
-        "agent_id" => agent.id,
-        "user_id" => user.id,
-        "prompt" => "hello"
-      }
-
-      assert {:ok, conv} = Conversations.start_conversation(attrs)
-      assert conv.agent_id == agent.id
-      assert conv.user_id == user.id
-      assert conv.status == "pending"
-    end
-
-    test "broadcasts graph update when parent_conversation_id is set", %{} do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      parent_conv = insert_conversation(user_id: user.id)
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _supervisor, _child_spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      attrs = %{
-        "agent_id" => agent.id,
-        "user_id" => user.id,
-        "parent_conversation_id" => parent_conv.id
-      }
-
-      assert {:ok, conv} = Conversations.start_conversation(attrs)
-      assert conv.parent_conversation_id == parent_conv.id
-    end
-
-    test "succeeds when vault_id is empty string", %{} do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _sup, _spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      attrs = %{"agent_id" => agent.id, "user_id" => user.id, "vault_id" => ""}
-      assert {:ok, conv} = Conversations.start_conversation(attrs)
-      assert is_nil(conv.vault_id)
-    end
-
-    test "succeeds and links vault when valid vault_id provided", %{} do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      vault = insert_vault(user_id: user.id)
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _sup, _spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      attrs = %{"agent_id" => agent.id, "user_id" => user.id, "vault_id" => vault.id}
-      assert {:ok, conv} = Conversations.start_conversation(attrs)
-      assert conv.vault_id == vault.id
-    end
-
-    test "allows a vault on the agent's allowed_vault_ids list", %{} do
-      user = insert_verified_user()
-      vault = insert_vault(user_id: user.id)
-      agent = insert_agent(user_id: user.id, allowed_vault_ids: [vault.id])
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _sup, _spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      attrs = %{"agent_id" => agent.id, "user_id" => user.id, "vault_id" => vault.id}
-      assert {:ok, conv} = Conversations.start_conversation(attrs)
-      assert conv.vault_id == vault.id
-    end
-
-    test "returns {:error, :vault_not_allowed} for a vault outside the allowlist", %{} do
-      user = insert_verified_user()
-      allowed = insert_vault(user_id: user.id)
-      other = insert_vault(user_id: user.id)
-      agent = insert_agent(user_id: user.id, allowed_vault_ids: [allowed.id])
-
-      attrs = %{"agent_id" => agent.id, "user_id" => user.id, "vault_id" => other.id}
-      assert {:error, :vault_not_allowed} = Conversations.start_conversation(attrs)
-    end
-
-    test "returns {:error, :vault_not_allowed} for any vault when the allowlist is empty", %{} do
-      user = insert_verified_user()
-      vault = insert_vault(user_id: user.id)
-      agent = insert_agent(user_id: user.id, allowed_vault_ids: [])
-
-      attrs = %{"agent_id" => agent.id, "user_id" => user.id, "vault_id" => vault.id}
-      assert {:error, :vault_not_allowed} = Conversations.start_conversation(attrs)
-    end
-
-    test "empty allowlist still permits starting with no vault at all", %{} do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id, allowed_vault_ids: [])
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _sup, _spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      attrs = %{"agent_id" => agent.id, "user_id" => user.id}
-      assert {:ok, conv} = Conversations.start_conversation(attrs)
-      assert is_nil(conv.vault_id)
-    end
-  end
-
-  # ────────────────────────────────────────────────────────────────────────────
-  # start_conversation/2 — per-launch environment override (#783)
-  # ────────────────────────────────────────────────────────────────────────────
-
-  describe "start_conversation/2 environment_id override" do
-    setup do
-      stub(Horde.DynamicSupervisor, :start_child, fn _sup, _spec ->
-        {:ok, spawn(fn -> :ok end)}
-      end)
-
-      user = insert_verified_user()
-      agent_env = insert_env(user_id: user.id)
-      other_env = insert_env(user_id: user.id)
-      agent = insert_agent(user_id: user.id, environment_id: agent_env.id)
-      %{user: user, agent: agent, agent_env: agent_env, other_env: other_env}
-    end
-
-    test "pins the conversation and its sandbox to the named environment", ctx do
-      attrs = %{
-        "agent_id" => ctx.agent.id,
-        "user_id" => ctx.user.id,
-        "environment_id" => ctx.other_env.id
-      }
-
-      assert {:ok, conv} = Conversations.start_conversation(attrs)
-      assert conv.environment_id == ctx.other_env.id
-
-      assert Conversations._unsafe_get_sandbox!(conv.sandbox_id).environment_id ==
-               ctx.other_env.id
-    end
-
-    test "nil and blank mean the agent's environment", ctx do
-      for value <- [nil, ""] do
-        attrs = %{
-          "agent_id" => ctx.agent.id,
-          "user_id" => ctx.user.id,
-          "environment_id" => value
-        }
-
-        assert {:ok, conv} = Conversations.start_conversation(attrs)
-        assert is_nil(conv.environment_id)
-
-        assert Conversations._unsafe_get_sandbox!(conv.sandbox_id).environment_id ==
-                 ctx.agent_env.id
-      end
-    end
-
-    test "a foreign environment reads as not found — same as an unknown id", ctx do
-      stranger = insert_verified_user()
-      foreign = insert_env(user_id: stranger.id)
-
-      for id <- [foreign.id, Ecto.UUID.generate()] do
-        attrs = %{"agent_id" => ctx.agent.id, "user_id" => ctx.user.id, "environment_id" => id}
-        assert {:error, :environment_not_found} = Conversations.start_conversation(attrs)
-      end
-    end
-
-    test "allowed_environment_ids: a listed environment passes, an unlisted one is refused",
-         ctx do
-      third = insert_env(user_id: ctx.user.id)
-
-      {:ok, agent} =
-        Agents.update_agent(ctx.agent, %{allowed_environment_ids: [ctx.other_env.id]})
-
-      ok = %{
-        "agent_id" => agent.id,
-        "user_id" => ctx.user.id,
-        "environment_id" => ctx.other_env.id
-      }
-
-      assert {:ok, conv} = Conversations.start_conversation(ok)
-      assert conv.environment_id == ctx.other_env.id
-
-      refused = %{"agent_id" => agent.id, "user_id" => ctx.user.id, "environment_id" => third.id}
-      assert {:error, :environment_not_allowed} = Conversations.start_conversation(refused)
-    end
-
-    test "an empty allowlist forbids every override but still permits the agent's own", ctx do
-      {:ok, agent} = Agents.update_agent(ctx.agent, %{allowed_environment_ids: []})
-
-      refused = %{
-        "agent_id" => agent.id,
-        "user_id" => ctx.user.id,
-        "environment_id" => ctx.other_env.id
-      }
-
-      assert {:error, :environment_not_allowed} = Conversations.start_conversation(refused)
-
-      # Naming the agent's own environment is not an override — it is pinned
-      # (a later change of the agent's environment does not move it), but it
-      # needs no allowlist entry.
-      own = %{
-        "agent_id" => agent.id,
-        "user_id" => ctx.user.id,
-        "environment_id" => ctx.agent_env.id
-      }
-
-      assert {:ok, conv} = Conversations.start_conversation(own)
-      assert conv.environment_id == ctx.agent_env.id
-    end
-
-    test "the allowlist is checked before the fetch, so a foreign id is refused not probed",
-         ctx do
-      stranger = insert_verified_user()
-      foreign = insert_env(user_id: stranger.id)
-      {:ok, agent} = Agents.update_agent(ctx.agent, %{allowed_environment_ids: []})
-
-      attrs = %{"agent_id" => agent.id, "user_id" => ctx.user.id, "environment_id" => foreign.id}
-      assert {:error, :environment_not_allowed} = Conversations.start_conversation(attrs)
-    end
-
-    test "channel resume keys on the override: a different environment is a different binding",
-         ctx do
-      base = %{"agent_id" => ctx.agent.id, "user_id" => ctx.user.id, "channel_id" => "chan-783"}
-
-      assert {:ok, first, :created} =
-               Conversations.start_or_resume_conversation(
-                 Map.put(base, "environment_id", ctx.other_env.id)
-               )
-
-      # Same channel, same environment: resumed.
-      assert {:ok, again, :resumed} =
-               Conversations.start_or_resume_conversation(
-                 Map.put(base, "environment_id", ctx.other_env.id)
-               )
-
-      assert again.id == first.id
-
-      # Same channel, no override: a fresh conversation, not the pinned one.
-      assert {:ok, plain, :created} = Conversations.start_or_resume_conversation(base)
-      refute plain.id == first.id
-      assert is_nil(plain.environment_id)
-
-      # And the plain one is now what a plain launch resumes.
-      assert {:ok, resumed, :resumed} = Conversations.start_or_resume_conversation(base)
-      assert resumed.id == plain.id
-    end
-  end
-
-  # ────────────────────────────────────────────────────────────────────────────
-  # _unsafe_list_active_conversations/0 — ordering
-  # ────────────────────────────────────────────────────────────────────────────
-
-  describe "_unsafe_list_active_conversations/0 ordering" do
-    test "running conversations appear before idle conversations" do
-      user = insert_verified_user()
-      idle = insert_conversation(user_id: user.id, status: "idle")
-      running = insert_conversation(user_id: user.id, status: "running")
-
-      results = Conversations._unsafe_list_active_conversations()
-      active_ids = results |> Enum.map(& &1.id) |> Enum.filter(&(&1 in [idle.id, running.id]))
-      assert hd(active_ids) == running.id
-    end
-  end
-
-  # ────────────────────────────────────────────────────────────────────────────
-  # _unsafe_list_turns/1 — image preloads
-  # ────────────────────────────────────────────────────────────────────────────
-
-  describe "_unsafe_list_turns/1 image preloads" do
-    test "preloads images association on each turn" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      %Fountain.Conversations.TurnImage{}
-      |> Fountain.Conversations.TurnImage.changeset(%{
-        turn_id: turn.id,
-        position: 0,
-        media_type: "image/png",
-        data: <<1, 2, 3>>
-      })
-      |> Ecto.Changeset.put_change(:inserted_at, DateTime.utc_now() |> DateTime.truncate(:second))
-      |> Repo.insert!()
-
-      [loaded_turn] = Conversations._unsafe_list_turns(conv.id)
-      assert length(loaded_turn.images) == 1
-      [img] = loaded_turn.images
-      assert img.media_type == "image/png"
-      assert img.data == <<1, 2, 3>>
-    end
-
-    test "returns turns with empty images list when no images have been inserted" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      _turn = insert_turn(conv)
-
-      [loaded_turn] = Conversations._unsafe_list_turns(conv.id)
-      assert loaded_turn.images == []
-    end
-
-    test "orders images by position ascending when multiple images exist" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      turn = insert_turn(conv)
-
-      now = DateTime.utc_now() |> DateTime.truncate(:second)
-
-      for {mt, data, pos} <- [
-            {"image/png", <<10>>, 0},
-            {"image/jpeg", <<20>>, 1},
-            {"image/gif", <<30>>, 2}
-          ] do
-        %Fountain.Conversations.TurnImage{}
-        |> Fountain.Conversations.TurnImage.changeset(%{
-          turn_id: turn.id,
-          position: pos,
-          media_type: mt,
-          data: data
-        })
-        |> Ecto.Changeset.put_change(:inserted_at, now)
-        |> Repo.insert!()
-      end
-
-      [loaded_turn] = Conversations._unsafe_list_turns(conv.id)
-      positions = Enum.map(loaded_turn.images, & &1.position)
-      assert positions == Enum.sort(positions)
-    end
-  end
-
-  describe "insert_sandbox/1 factory — no explicit user_id" do
-    test "creates a new user when no user_id is provided" do
-      # Triggers the insert_verified_user().id fallback in insert_sandbox (factory.ex line 129)
-      sandbox = insert_sandbox()
-      assert is_binary(sandbox.user_id)
-    end
-  end
-
-  describe "insert_conversation/1 factory — agent without explicit user_id" do
-    test "derives user_id from agent when no user_id is provided" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      conv = insert_conversation(agent: agent)
-      assert conv.user_id == user.id
-      assert conv.agent_id == agent.id
-    end
-
-    test "creates a new user when neither user_id nor agent is provided" do
-      # Triggers the insert_verified_user().id fallback (factory.ex line 147)
-      conv = insert_conversation()
-      assert is_binary(conv.user_id)
-      assert is_nil(conv.agent_id)
-    end
-  end
-
-  describe "factory to_atom_map — safe_to_existing_atom fallback" do
-    test "to_atom_map with an unknown string key does not crash and returns the string as fallback" do
-      # \"xyzquuxfoo_novel_key_never_an_atom\" is not an existing Elixir atom,
-      # so safe_to_existing_atom triggers its rescue clause and returns the string key.
-      result =
-        Fountain.Factory.to_atom_map(%{
-          "sprite_name" => "test-sprite",
-          "xyzquuxfoo_novel_key_never_an_atom" => "ignored_value"
-        })
-
-      assert Map.get(result, :sprite_name) == "test-sprite"
-      # The unknown key is preserved as a string (fallback)
-      assert Map.get(result, "xyzquuxfoo_novel_key_never_an_atom") == "ignored_value"
-    end
-  end
-
-  # ────────────────────────────────────────────────────────────────────────────
-  # mark_read/2
-  # ────────────────────────────────────────────────────────────────────────────
-
-  describe "mark_read/2" do
-    test "sets last_read_at to approximately now" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      assert is_nil(conv.last_read_at)
-
-      before = DateTime.utc_now()
-      :ok = Conversations.mark_read(conv.id, user.id)
-      after_time = DateTime.utc_now()
-
-      reloaded = Conversations.get_conversation(conv.id, user.id)
-      refute is_nil(reloaded.last_read_at)
-      assert DateTime.compare(reloaded.last_read_at, before) in [:gt, :eq]
-      assert DateTime.compare(reloaded.last_read_at, after_time) in [:lt, :eq]
-    end
-
-    test "is scoped to owner — does not update another user's conversation" do
-      user1 = insert_verified_user()
-      user2 = insert_verified_user()
-      conv = insert_conversation(user_id: user1.id)
-
-      :ok = Conversations.mark_read(conv.id, user2.id)
-
-      reloaded = Conversations.get_conversation(conv.id, user1.id)
-      assert is_nil(reloaded.last_read_at)
-    end
-
-    test "calling twice updates last_read_at to a more recent time" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-
-      :ok = Conversations.mark_read(conv.id, user.id)
-      first_read = Conversations.get_conversation(conv.id, user.id).last_read_at
-
-      Process.sleep(2)
-
-      :ok = Conversations.mark_read(conv.id, user.id)
-      second_read = Conversations.get_conversation(conv.id, user.id).last_read_at
-
-      assert DateTime.compare(second_read, first_read) in [:gt, :eq]
-    end
-  end
-
-  # ────────────────────────────────────────────────────────────────────────────
-  # list_conversations/2 — last_active_at and unread safety
-  # ────────────────────────────────────────────────────────────────────────────
-
-  describe "list_conversations/2 last_active_at" do
-    test "defaults to inserted_at when there are no log events" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-
-      [result] = Conversations.list_conversations(user.id)
-
-      assert DateTime.compare(
-               result.last_active_at,
-               DateTime.from_naive!(result.inserted_at, "Etc/UTC")
-             ) == :eq
-    end
-
-    test "advances past inserted_at when output log events exist" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-
-      later = DateTime.utc_now() |> DateTime.add(60, :second) |> DateTime.truncate(:microsecond)
-      insert_log_event(conv, kind: "output", stream: "stdout", inserted_at: later)
-
-      [result] = Conversations.list_conversations(user.id)
-
-      assert DateTime.compare(
-               result.last_active_at,
-               DateTime.from_naive!(conv.inserted_at, "Etc/UTC")
-             ) == :gt
-    end
-
-    test "stage events (reconnects) do NOT advance last_active_at" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-
-      baseline_at = List.first(Conversations.list_conversations(user.id)).last_active_at
-
-      later = DateTime.utc_now() |> DateTime.add(300, :second) |> DateTime.truncate(:microsecond)
-      insert_log_event(conv, kind: "stage", data: "reattach", stream: "", inserted_at: later)
-
-      [result] = Conversations.list_conversations(user.id)
-      assert DateTime.compare(result.last_active_at, baseline_at) == :eq
-    end
-
-    test "last_read_at is nil by default" do
-      user = insert_verified_user()
-      insert_conversation(user_id: user.id)
-
-      [result] = Conversations.list_conversations(user.id)
-      assert is_nil(result.last_read_at)
-    end
-
-    test "last_read_at is populated after mark_read" do
-      user = insert_verified_user()
-      conv = insert_conversation(user_id: user.id)
-      Conversations.mark_read(conv.id, user.id)
-
-      [result] = Conversations.list_conversations(user.id)
-      refute is_nil(result.last_read_at)
-    end
-  end
-
-  describe "_unsafe_reap_sandbox/1" do
-    test "marks a ready sandbox with no live server terminated; conversations untouched" do
-      sandbox = insert_sandbox(status: "ready")
-      conv = insert_conversation(sandbox: sandbox, user_id: sandbox.user_id)
-
-      assert {:ok, :released} = Conversations._unsafe_reap_sandbox(sandbox.id)
-
-      reloaded = Conversations._unsafe_get_sandbox!(sandbox.id)
-      assert reloaded.status == "terminated"
-      assert reloaded.terminated_at
-      refute Conversations._unsafe_get_conversation(conv.id).status == "terminated"
-    end
-
-    test "is a no-op on an already-terminal sandbox" do
-      sandbox = insert_sandbox(status: "terminated")
-      assert {:ok, :already_terminal} = Conversations._unsafe_reap_sandbox(sandbox.id)
-    end
-
-    test "returns not_found for an unknown id" do
-      assert {:error, :not_found} = Conversations._unsafe_reap_sandbox(Ecto.UUID.generate())
-    end
-  end
-
-  describe "concurrent wake loses cleanly (#330)" do
-    # Two prompts race to wake the same dormant conversation. Both used to
-    # reach create_fresh_sandbox_and_start; the loser's start_child returned
-    # {:error, {:already_started, _}}, the `with` fell through, and its
-    # just-created pending sandbox row sat holding a quota slot until the
-    # reaper's pass an hour later — a user at their cap could lock themselves
-    # out for an hour by double-clicking.
-
-    test "the loser terminates its own sandbox row instead of stranding it" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      conv = insert_conversation(user_id: user.id, agent: agent, status: "idle")
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _sup, _spec ->
-        {:error, {:already_started, self()}}
-      end)
-
-      stub(Fountain.Conversations.ConversationServer, :queue_initial_prompt, fn _conv_id,
-                                                                                _prompt ->
-        :ok
-      end)
-
-      before = Fountain.Repo.all(Fountain.Conversations.Sandbox) |> MapSet.new(& &1.id)
-
-      assert {:ok, _conv} = Conversations.wake_conversation(conv.id, "hi")
-
-      # The row the loser created for itself is terminated, not pending —
-      # which is what #330 is about: a double-click must not strand a quota
-      # slot until the reaper's next pass.
-      created =
-        Fountain.Conversations.Sandbox
-        |> Fountain.Repo.all()
-        |> Enum.reject(&MapSet.member?(before, &1.id))
-
-      assert created != []
-
-      for sandbox <- created do
-        assert sandbox.status == "terminated"
-      end
-
-      # The conversation's *existing* sandbox is deliberately left alone now
-      # (#717). The loser does not know what the winner is running on, and
-      # `wake_conversation`'s reuse arm starts a server against exactly this
-      # sandbox — so retiring it here could terminate the one the winner is
-      # using. The winner retires it if it provisioned a replacement.
-      assert Fountain.Quotas.active_sandbox_count(user.id) == 1
-    end
-
-    test "the loser forwards its prompt to the winner" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      conv = insert_conversation(user_id: user.id, agent: agent, status: "idle")
-      test_pid = self()
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _sup, _spec ->
-        {:error, {:already_started, self()}}
-      end)
-
-      stub(Fountain.Conversations.ConversationServer, :queue_initial_prompt, fn conv_id, prompt ->
-        send(test_pid, {:forwarded, conv_id, prompt})
-        :ok
-      end)
-
-      assert {:ok, _} = Conversations.wake_conversation(conv.id, "the racing prompt")
-      assert_received {:forwarded, _conv_id, "the racing prompt"}
-    end
-
-    test "a genuine start failure still surfaces as an error" do
-      user = insert_verified_user()
-      agent = insert_agent(user_id: user.id)
-      conv = insert_conversation(user_id: user.id, agent: agent, status: "idle")
-
-      stub(Horde.DynamicSupervisor, :start_child, fn _sup, _spec ->
-        {:error, :max_children}
-      end)
-
-      assert {:error, :max_children} = Conversations.wake_conversation(conv.id, "hi")
-    end
-  end
 end
